feat(sizing): correctness fixes behind --sizing-v2; futures/QC/Quantpedia queue
Sizing v2 (opt-in via SizingPolicy; monitor sizes exactly as before without one):
- risk/quantity.py QuantityRule(step, min_qty, min_notional) floors to a venue's real
  lot instead of whole units, so a 0.36 PAXG position (~$1,560) is no longer swallowed to 0.
- KrakenBroker.lot_rules() reads lot_decimals/ordermin/costmin from public AssetPairs;
  to_kraken_pair now drops the slash for unlisted pairs (SOL/ADA/POL/UNI/AAVE/ZEC matched).
- risk/sizing_policy.py: per-venue annualization (365 Kraken, 252 QT/IB), enforced ATR
  stops (rebuilt from avg entry after a restart), and a state/sizing_decisions.jsonl journal
  recording every conviction factor and the reason a size rounded to zero.
- Router.quantity_rule_for trims size-cap breaches to the same lots; OrderIntent.shares,
  SizingResult.shares, approval/notification quantities are now float.
- Monitor exit closes the exact held quantity (always on) so a fractional crypto position
  can be sold. scripts/paper_kraken.py --sizing-v2 wires it (refuses to launch on a missing
  lot rule). tests/test_sizing_v2.py.

Also queued from 2026-09-13 (see NEXT_SESSION.md): IB-trading-class futures specs, futures
discovery + global book wired to the intel graph, LEAN/QC futures mirror + empirical runner,
Quantpedia API client and CLI. Reports for those sessions included.

+++ src/trading_live_claude/execution/approval.py
@@ -103,7 +103,7 @@ class Prompt:
     broker: str            # "ib" | "kraken" | "questrade" — destination brokerage
     symbol: str
     action: str
-    shares: int
+    shares: float
     entry: float
     stop: float
     target: float | None
@@ -160,7 +160,7 @@ class PassbookEntry:
     broker: str
     symbol: str
     action: str
-    shares: int
+    shares: float
     notional_usd: float
     strategy: str
     thesis: str
